requirements/doc.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
agate==1.7.1
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   dbt-snowflake
alabaster==0.7.13
    # via sphinx
annotated-types==0.6.0
    # via
    #   -r requirements/test.txt
    #   pydantic
asn1crypto==1.5.1
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
attrs==23.2.0
    # via
    #   -r requirements/test.txt
    #   jsonschema
    #   referencing
babel==2.14.0
    # via
    #   -r requirements/test.txt
    #   agate
    #   sphinx
certifi==2024.2.2
    # via
    #   -r requirements/test.txt
    #   requests
    #   snowflake-connector-python
cffi==1.16.0
    # via
    #   -r requirements/test.txt
    #   cryptography
    #   dbt-core
    #   snowflake-connector-python
charset-normalizer==3.3.2
    # via
    #   -r requirements/test.txt
    #   requests
    #   snowflake-connector-python
click==8.1.7
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   dbt-semantic-interfaces
colorama==0.4.6
    # via
    #   -r requirements/test.txt
    #   dbt-core
coverage[toml]==7.4.3
    # via
    #   -r requirements/test.txt
    #   pytest-cov
cryptography==42.0.5
    # via
    #   -r requirements/test.txt
    #   pyopenssl
    #   snowflake-connector-python
dbt-core==1.7.10
    # via
    #   -r requirements/test.txt
    #   dbt-snowflake
dbt-extractor==0.5.1
    # via
    #   -r requirements/test.txt
    #   dbt-core
dbt-semantic-interfaces==0.4.4
    # via
    #   -r requirements/test.txt
    #   dbt-core
dbt-snowflake==1.7.2
    # via -r requirements/test.txt
doc8==1.1.1
    # via -r requirements/doc.in
docutils==0.19
    # via
    #   doc8
    #   readme-renderer
    #   restructuredtext-lint
    #   sphinx
edx-sphinx-theme==3.1.0
    # via -r requirements/doc.in
exceptiongroup==1.2.0
    # via
    #   -r requirements/test.txt
    #   pytest
filelock==3.13.1
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
idna==3.6
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   requests
    #   snowflake-connector-python
imagesize==1.4.1
    # via sphinx
importlib-metadata==6.11.0
    # via
    #   -r requirements/test.txt
    #   dbt-semantic-interfaces
    #   keyring
    #   sphinx
importlib-resources==6.3.0
    # via
    #   -r requirements/test.txt
    #   jsonschema
    #   jsonschema-specifications
    #   keyring
iniconfig==2.0.0
    # via
    #   -r requirements/test.txt
    #   pytest
isodate==0.6.1
    # via
    #   -r requirements/test.txt
    #   agate
    #   dbt-core
jaraco-classes==3.3.1
    # via
    #   -r requirements/test.txt
    #   keyring
jinja2==3.1.3
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   dbt-semantic-interfaces
    #   sphinx
jsonschema==4.21.1
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   dbt-semantic-interfaces
jsonschema-specifications==2023.12.1
    # via
    #   -r requirements/test.txt
    #   jsonschema
keyring==24.3.1
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
leather==0.4.0
    # via
    #   -r requirements/test.txt
    #   agate
logbook==1.5.3
    # via
    #   -r requirements/test.txt
    #   dbt-core
markupsafe==2.1.5
    # via
    #   -r requirements/test.txt
    #   jinja2
mashumaro[msgpack]==3.12
    # via
    #   -r requirements/test.txt
    #   dbt-core
minimal-snowplow-tracker==0.0.2
    # via
    #   -r requirements/test.txt
    #   dbt-core
more-itertools==10.2.0
    # via
    #   -r requirements/test.txt
    #   dbt-semantic-interfaces
    #   jaraco-classes
msgpack==1.0.8
    # via
    #   -r requirements/test.txt
    #   mashumaro
networkx==3.1
    # via
    #   -r requirements/test.txt
    #   dbt-core
nh3==0.2.15
    # via readme-renderer
packaging==23.1
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/test.txt
    #   dbt-core
    #   pytest
    #   snowflake-connector-python
    #   sphinx
parsedatetime==2.6
    # via
    #   -r requirements/test.txt
    #   agate
pathspec==0.11.2
    # via
    #   -r requirements/test.txt
    #   dbt-core
pbr==6.0.0
    # via stevedore
pkgutil-resolve-name==1.3.10
    # via
    #   -r requirements/test.txt
    #   jsonschema
platformdirs==3.11.0
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
pluggy==1.4.0
    # via
    #   -r requirements/test.txt
    #   pytest
protobuf==4.25.3
    # via
    #   -r requirements/test.txt
    #   dbt-core
pycparser==2.21
    # via
    #   -r requirements/test.txt
    #   cffi
pydantic==2.6.4
    # via
    #   -r requirements/test.txt
    #   dbt-semantic-interfaces
pydantic-core==2.16.3
    # via
    #   -r requirements/test.txt
    #   pydantic
pygments==2.17.2
    # via
    #   doc8
    #   readme-renderer
    #   sphinx
pyjwt==2.8.0
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
pyopenssl==24.1.0
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
pytest==8.1.1
    # via
    #   -r requirements/test.txt
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements/test.txt
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/test.txt
    #   dbt-semantic-interfaces
python-slugify==8.0.4
    # via
    #   -r requirements/test.txt
    #   agate
pytimeparse==1.1.8
    # via
    #   -r requirements/test.txt
    #   agate
pytz==2024.1
    # via
    #   -r requirements/test.txt
    #   babel
    #   dbt-core
    #   snowflake-connector-python
pyyaml==6.0.1
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   dbt-semantic-interfaces
readme-renderer==43.0
    # via -r requirements/doc.in
referencing==0.33.0
    # via
    #   -r requirements/test.txt
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   minimal-snowplow-tracker
    #   snowflake-connector-python
    #   sphinx
restructuredtext-lint==1.4.0
    # via doc8
rpds-py==0.18.0
    # via
    #   -r requirements/test.txt
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   -r requirements/test.txt
    #   edx-sphinx-theme
    #   isodate
    #   minimal-snowplow-tracker
    #   python-dateutil
snowballstemmer==2.2.0
    # via sphinx
snowflake-connector-python[secure-local-storage]==3.7.1
    # via
    #   -r requirements/test.txt
    #   dbt-snowflake
sortedcontainers==2.4.0
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
sphinx==5.3.0
    # via
    #   -r requirements/doc.in
    #   edx-sphinx-theme
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.4
    # via
    #   -r requirements/test.txt
    #   dbt-core
stevedore==5.2.0
    # via doc8
text-unidecode==1.3
    # via
    #   -r requirements/test.txt
    #   python-slugify
tomli==2.0.1
    # via
    #   -r requirements/test.txt
    #   coverage
    #   doc8
    #   pytest
tomlkit==0.12.4
    # via
    #   -r requirements/test.txt
    #   snowflake-connector-python
typing-extensions==4.10.0
    # via
    #   -r requirements/test.txt
    #   annotated-types
    #   dbt-core
    #   dbt-semantic-interfaces
    #   mashumaro
    #   pydantic
    #   pydantic-core
    #   snowflake-connector-python
urllib3==1.26.18
    # via
    #   -r requirements/test.txt
    #   dbt-core
    #   requests
    #   snowflake-connector-python
zipp==3.18.0
    # via
    #   -r requirements/test.txt
    #   importlib-metadata
    #   importlib-resources
